subscription activation added to checkout form
