+++ src/Pages/UserSubscription/UserSubscription.jsx
@@ -3,32 +3,16 @@ import { loadStripe } from "@stripe/stripe-js";
 import CheckoutForm from "./CheckoutForm";
 import payment from "../../assets/payment.json"
 import Lottie from "lottie-react";
-import { useState } from "react";
 import "./UserSubscription.css"
 const stripePromise = loadStripe(import.meta.env.VITE_PK)
 const UserSubscription = () => {
-    const [activeone,setActiveone] = useState(true);
-    const [activetwo,setActivetwo] = useState(false);
-
-    const activeMonthly = () =>{
-        setActiveone(true);
-        setActivetwo(false);
-    }
-    const activeYearly = () =>{
-        setActiveone(false);
-        setActivetwo(true);
-    }
+    
     return (
         <div className="lg:flex gap-10">
             <div className='w-1/2 py-10 px-5 mx-auto'>
                 <h1 className="font-bold font-playfair text-4xl text-center">Weave<span className="text-teal">PDF</span></h1><br />
                 <p className="font-bold font-playfair text-lg md:text-xl text-center">Get premium to enjoy endless exciting features</p><br /><br />
-                <div className="flex gap-10 justify-center">
-                    <div onClick={activeMonthly} className={activeone? 'activeButton' : 'deactiveButton'}>
-                        Monthly <br />$ 50</div>
-                    <div onClick={activeYearly} className={activetwo? 'activeButton' : 'deactiveButton'}>
-                        Yearly <br />$ 540</div>
-                    </div><br /><br />
+                
                 <Elements stripe={stripePromise}>
                     <CheckoutForm></CheckoutForm>
                 </Elements>
